twitter_clone_grpc/requirements.txt
psycopg2-binary>=2.8
alembic>=1.4.3
grpcio==1.36.1
grpcio-tools==1.36.1
SQLAlchemy>=1.3.20
pytest>=6.1.1